[not verified] Pass the login string to UI
