+++ projects/plugins/jetpack/extensions/shared/memberships.js
@@ -48,3 +48,34 @@ export const initializeMembershipButtons = selector => {
 		button.setAttribute( 'data-jetpack-memberships-button-initialized', 'true' );
 	} );
 };
+
+const updateQueryStringParameter = function ( uri, key, value ) {
+	const re = new RegExp( '([?&])' + key + '=.*?(&|$)', 'i' );
+	const separator = uri.indexOf( '?' ) !== -1 ? '&' : '?';
+	if ( uri.match( re ) ) {
+		return uri.replace( re, '$1' + key + '=' + value + '$2' );
+	}
+	return uri + separator + key + '=' + value;
+};
+
+export const setPurchaseResultCookie = function ( premiumContentJWTToken ) {
+	// We will set this in a cookie  - just in case. This will be reloaded in the refresh, when user clicks OK.
+	// But user can close the browser window before clicking OK. IN that case, we want to leave a cookie behind.
+	const date = new Date();
+	date.setTime( date.getTime() + 365 * 24 * 60 * 60 * 1000 );
+	document.cookie =
+		'jp-premium-content-session' +
+		'=' +
+		premiumContentJWTToken +
+		'; expires=' +
+		date.toGMTString() +
+		'; path=/';
+};
+
+export const reloadPageWithPremiumContentQueryString = function ( premiumContentJWTToken ) {
+	document.location.href = updateQueryStringParameter(
+		document.location.href,
+		'token',
+		premiumContentJWTToken
+	);
+};
